[shardformer] refactored the shardformer layer structure (#4053)
