+++ colossalai/shardformer/layer/__init__.py
@@ -1,17 +1,10 @@
 from .dropout import Dropout1D
-from .embedding1d import Embedding1D
-from .layernorm1d import LayerNorm1D
-from .linear1d import Linear1D_Col, Linear1D_Row
-from .linearconv1d import LinearConv1D_Col, LinearConv1D_Row
-from .vocabparallelembedding1d import VocabParallelEmbedding1D
+from .embedding import Embedding1D, VocabParallelEmbedding1D
+from .linear import Linear1D_Col, Linear1D_Row
+from .linear_conv import LinearConv1D_Col, LinearConv1D_Row
+from .loss import cross_entropy_1d
 
 __all__ = [
-    "Embedding1D",
-    "VocabParallelEmbedding1D",
-    "Linear1D_Col",
-    "Linear1D_Row",
-    "LinearConv1D_Col",
-    "LinearConv1D_Row",
-    "LayerNorm1D",
-    "Dropout1D",
+    "Embedding1D", "VocabParallelEmbedding1D", "Linear1D_Col", "Linear1D_Row", "LinearConv1D_Col", "LinearConv1D_Row",
+    "Dropout1D", "cross_entropy_1d"
 ]
