+++ colossalai/shardformer/layer/_operation.py
@@ -1,8 +1,6 @@
 import torch
 import torch.distributed as dist
 
-from colossalai.core import global_context as gpc
-
 try:
     import fused_mix_prec_layer_norm_cuda
 except:
