+++ tests/test_shardformer/test_layer/test_embedding.py
@@ -4,7 +4,7 @@ import torch.nn as nn
 from torch.testing import assert_close
 
 import colossalai
-from colossalai.shardformer.layer.layers import Embedding1D
+from colossalai.shardformer.layer import Embedding1D
 from colossalai.testing import rerun_if_address_is_in_use, spawn
